Fix for possible hanging go routines in error conditions
`client.go` - SyncProviderConfig no longer waits for the first sync to complete.  This is important because if the first sync never completes the calling function has no way to interrupt or cancel the attempts to sync.  This may result in go routines forever looping if an invalid configuration was provided.

Calling functions that want to block till the first sync must now do so through a new returned function pointer.  This API change is unfortunate but I saw no way to fix the root problem without putting the calling function in control of the blocking behavior.

`provider.go` - If the stop channel is invoked then this will be considered the "first sync".  This way any go routines blocked waiting for the first sync will be woken up if the attempts to sync is canceled.